[HTTP] Fix new diagnostic test (#117248)
Depending on the timing of the cancellation, the exception type can be either `OperationCanceledException` or `TaskCanceledExcption`, we need to allow both. The PR also refactors the assertions outside of the `ActivityListener` callback, so we get better error messages if something fails in the future. (This test has been recently added in #116269.)