Remove custom enumerator from ConcurrentBag
ConcurrentBag's snapshot enumeration semantics are obtained by creating an array of the contents and then enumerating that array. This was done with a custom enumerator out of an abundance of caution about maintaining semantics for operations with undefined behavior, e.g. using Current after MoveNext returns false. We've since stopped trying to maintain bug-for-bug compatibility in such undefined areas. As such, this just deletes the custom enumerator and uses Array's. This not only reduces code but has better performance characteristics (though the whole ToArray as part of enumeration already dominates the enumeration costs).