Fix IP adjustment for interpreter EH (#117055)
* Fix IP adjustment for interpreter EH

The exception handling stack frame iterator adjusts IP it returns for
all cases when the IP represents a return address. This is to make sure
that if the last instruction in a try regios is a call, it is still covered
by the try region (the return address would be right after it).
For jit/aoted code, throwing a software exceptions results in a call.
Only hardware exceptions like division by zero etc report the address of
the failing instruction.
In the interpreter, when an exception is thrown, the IP address is
always the address of the instruction that has triggered the exception.
So we should not adjust the IP.
This change fixes that by marking exception throwing frame as
"faulting", which ensures the adjustment doesn't occur.
That information is recorded in the InterpreterFrame when an
exception is thrown by the interpreter code and cleared when the
execution resumes after catch.

* Fix build break with interpreter disabled

* Fix the case when the pEntryFrame is NULL

* Actually, the check should be for FRAME_TOP, not NULL