JIT: relax another check in conditional escape analysis (#117295)
Proceed with CEA cloning even when the allocation site dominates the assignment
to the enumerator var. The cloning will be a bit wasteful as the original code will
become unreachable, but the code is set up to specialize the clone and so this is
the less risky fix.

Also restore recognition of `IEnumerable<T>.GetEnumerator` as this gives a useful
inlining boost in some cases.

Fixes #117204 cases that were not fixed by #117222.