JIT: use local zero-init for stack objects in more cases
We sometimes see poor CQ after stack allocating and promotion, because
we have decided to zero-init in the prolog and this makes the promoted
fields be EH live.

Try and mitigate this by using local zero-init if we see that we may
pass through EH before reaching the allocation site. To determine this
we use a simplistic check of the blocks in RPO that appear before the
allocation block; if any is a try entry, we do local zero-init.