Fix FrameworkDependentAppLaunch.AppHost_GlobalLocation 'Failed to restore file' CI failure (#116831)
We're hitting an intermittent failure (particularly on Debug) trying to copy over files immediately after the process exits. Since this passes the majority of the time, it does not seem to indicate an actual issue where the files are left in use.
- Increase the existing retry we have for doing this restore
- Move tests to InstallLocation test class as it fits better there (testing registered and default install locations)
- Make InstallLocation test class create one app with test behaviour enabled instead of repeatedly doing it in half of its tests cases.

This also removes explicitly running the app with the app directory as the working directory, which doesn't seem interesting for us to keep testing.