Ensure profiler test output is flushed
This might fix issue #114608

These tests redirect profilee output to a parent which can cause it to get buffered. Given that the last line of output is
missing but the Profiler.dll!Profiler::Shutdown message emitted just before is present it seems more likely that this
is a buffering issue rather than a process crash in that small window. Notably, every other profiler Shutdown() method
includes this fflush() call but it was missing for the transitions profiler.