Fix args passed by stack ref in the CallStubGenerator (#117399)
* Fix args passed by stack ref in the CallStubGenerator

The CallStubGenerator was missing support for arguments passed by
reference using stack slots. Only references in registers were
implemented. There was also a bug in passing return buffer argument when
there was "this" at the same time. Only the m_r1 was set to 1, but the
m_r2 was left unchanged, being 0. That resulted in the routine fetched
for that register range being NULL.

This change fixes both of these issues and adds a test.

* Fix arm64 linux asm break due to accidental semicolon