Add InlinedCallFrame support for the interpreter take 2
This change adds support for using InlinedCallFrame in the interpreter.
It adds a public constructor and also ensures that the stack walking
works correctly. The stack walking is tricky, because the the runtime
expects the InlinedCallFrame updates REGDISPLAY to point to managed
code. In the interpreter case, that means the interpreted frames. But
that changes how the stack walking enters the interpreted frames.
Without the InlinedCallFrame, the stack walking sees that it would move
to the InterpreterFrame and sets the REGDISPLAY to the interpreted
context and enumerates all the interpreted frames for the
InterpreterFrame. Once all the frames are enumerated, it moves to the
InterpretedFrame itself and let's it update the REGDISPLAY to the caller
of the interpreted code.
With the InlinedCallFrame, we move to the interpreted frames before we
see the InterpreterFrame. And we also need to ensure that the first
argument register of the interpreter context is set to point to the
InterpreterFrame for the stack unwinding to work correctly.