Enable feature stubprecode dynamic helpers again and fix memory ordering logic around mapped stubs (#115746)
Use flush instruction cache to make this safe. Adding an address dependency got into lots of details around possibly needing an entire new stub type, and several paths through this code already depend on the concept of having a barrier here. To support modifying the stubs without inserting lots of magic number, rework how stubs are identified to just do a complete compare of the entire assembly stub in a way which will be easily to generalize to the cDAC.

In addition, make changes to make the FixupPrecode path safe from the same memory ordering concerns. Since the FixupPrecode always goes through a specified entrypoint, we can slide the barrier into only the slow path as long as we ensure that the Target field written with a VolatileStore after the MethodDesc and PrecodeFixupThunk fields are set. To ensure that the barrier on the writer side has happened before with hit the barrier on the reader, we ensure that the Target field is initialized into an infinite loop before the code associated with the FixupPrecode is mapped into the process.

Fixes #113810

Also, this is now only enabled for scenarios where it is required. (Mac Catalyst and iOS), as there was evidence of performance loss when examined with @EgorBot.