Ensure proper InterpThreadContext initialization (#117479)
* Ensure proper InterpThreadContext initialization

In case an interpreted method was compiled on one thread and then later
executed on another thread that never executed any interpreted code
before, the InterpThreadContext was not initialized.
This change fixes that by calling Thread::GetInterpThreadContext from
the InterpreterStub in case it was not initialized for the current
thread yet.

* Fix Unix Amd64