Fix TryInvokeICustomQueryInterface exception handling (#117518)
* Fix TryInvokeICustomQueryInterface exception handling

When the InteropLibImports::TryInvokeICustomQueryInterface is called
from a native thread and a managed exception occurs in the managed
ICustomQueryInterface interface implementation, the exception handling
reports that exception as unhandled because it doesn't know tha the
InteropLibImports::TryInvokeICustomQueryInterface catches managed
exceptions and converts them to HRESULTs.
This change fixes it by adding DebuggerU2MCatchHandlerFrame to the
InteropLibImports::TryInvokeICustomQueryInterface. That allows EH to
understand that the exception that flows to native code will be caught
and so it should not report it as unhandled.

Close #117393

* Add regression test

* Fix unix test build

* Fix unix test build again