Change RhNewString length argument to nint/intptr_t (#117552)
The RhNewString length argument is declared as int, but the asm
implementation was using the full 64 bits of the related argument
register. That doesn't match the calling convention where there is no
guarantee that the upper 32 bits are zeroed. That has caused problem in
the interpreter that always loads full registers from the interpreter
stack and so sometimes, the upper 32 bits were a garbage and the
function failed because the length was seemingly too large.

This change fixes it by changing signature of the RhNewString so that
the length is nint in managed code / intptr_t in native code.