Change unhandled exception detection
Currently, the code detecting whether an exception should be propagated
into native frames and not reported as unhandled relies on presence of
the DebuggerU2MFrame with "CatchAllExceptions" set. Recent fixes due to
this marking missing causing problems revealed that it would be actually
better to do the detection in an inverse manner. That means always
propagate the exception to native frames unless there is an explicit
marker indicating that the exception needs to be reported as unhandled.
Such a marker should only mark the main entry point for an application
and a managed thread entry. All other places should just let the
exception flow into the native code.

This change fixes it by creating a new kind of explicit frame to mark
the code paths that should report unhandled exception.