Fix ExecutionManager::GetCodeMethodDesc for interpreter
The ExecutionManager::GetCodeMethodDesc expects that the address
passed to it is an address of actual JIT/AOT or interpreter generated
code. However, it is sometimes passed an address of the interpreter
precode, e.g. in the case when called by the
MethodTable::GetMethodDescForSlotAddress. The VTable slots need to point
to the interpreter stub so that these methods can be called by virtual
calls from native code.

This change adds a check for the interpreter precode to the
ExecutionManager::GetCodeMethodDesc and extracts the MethodDesc from the
interpreter data in case it is passed the precode address.