Address COM interop case involving legacy `wCode` usage. (#117596)
* Address COM interop case involving legacy wCode usage.

The wCode field in the EXCEPINFO is a 16-bit signed integer. This
is a legacy field for 16-bit Windows, but is still used in many COM
servers. The current implementation assumed it would result in an
Exception using Marshal.GetExceptionForHR(), but converting a
short to an int, will rarely result in a negative value which means
null will almost always be returned. This change checks if the
exception is null and if so, creates a simple COMException with the
error code.