Drop usage of `.All` in Microsoft.Extensions
The existing usage is a leftover from when we didn't have the binding source generation and the guidance was to suppress warning and preserve everything manually. We had a similar thing for S.T.Json in the past that we also deleted/dropped in favor of the source generation.

All the updated places are already marked RUC.