Fix interpreter passing managed "this" in presence of return buffer (#117661)
The CallStubGenerator is using the native x64 Windows calling
convention for the case when both "this" and "return buffer" arguments
are passed to native callees. However, the CLR ABI diverges from this
convention. While the native one passes the return buffer first and the
this pointer second, the CLR ABI always passes the "this" pointer first.

This fixes a number of failures on the runtime startup path when running
interpreted.