JIT: Store unique switch successors in `BBswtDesc` (#117423)
* Rename `BBehfDesc` to `BBJumpTable`, and repurpose it for describing blocks with `N` successors. This type allows unique successor access, enumeration, and removal.
* Have `BBswtDesc` inherit from `BBJumpTable`. Through `BBswtDesc`'s updated API surface, we can enumerate unique successors and switch cases.
* Remove `Compiler`'s API surface for mapping switch blocks to their unique successors.

There are numerous points in the JIT where we want to iterate unique successors, but we tolerate iterating duplicates out of fear of needing to construct the switch successor map. We can now iterate unique successors at those sites without needing to precompute anything. I'll do that switchover in a follow-up PR.