Fix NativeAOT COM custom query interface exception handling (#117691)
* Fix NativeAOT COM custom query interface exceptioon handling

After fixing a bug in the related stuff in coreclr, the regression test
has revealed that the NativeAOT version has a similar issue. The
exception stemming from the custom query interface are not caught
and transformed to HRESULT, so they are reported as unhandled
exceptions.

This change adds try / catch to catch the exception and translate it to
HRESULT.

* PR feedback