Forward CancellationToken in Task-based Socket.ConnectAsync (#116943)
Fixes #75889

Unlike other Task-based Socket operations, `ConnectAsync` does not forward the `CancelationToken` to the PAL, instead it creates a CTR that uses `CancelConnectAsync` which is implemented by simply disposing the socket. This can lead to a race between Connect completion and socket disposal resulting in an invalid socket state where a disposed or non-connected socket is returned from a seemingly succesful connect attempt.

I couldn't find a way to implement a deterministic test for the scenario in #75889, but according to my local testing, the issue is gone with the changes.