[HTTP Telemetry] Prioritize Host header when emitting `server.address` (#117540)
With the merge of https://github.com/open-telemetry/semantic-conventions/pull/2463, we are good to prioritize the contents of the `Host` header in cases no proxy is being used. The PR implements the change for both request and connection traces+metrics.

There is a non-negligible risk: actually, we do not and (with the current code structure) can not emit `network.peer.address` for request telemetry, meaning that with `http(s)://x.x.x.x/..` target Uri-s, IP information will be no longer present when a there is a Host header. I believe that most users would still prefer to see the contents of the Host header if there is a mismatch. Others can opt into connection metrics/traces where `network.peer.address` is available.