JIT: constrain max IL size for OSR inlining
If we inline too aggressively at intermediate tiers, we can lose profile
data that would be beneficial at the final tier.

Extend the recent fixes made for limiting max IL size at Tier1+Instr to
include OSR as well. This one is a arguably bit more delicate as OSR may
well be the final tier, but there is no way to know that.

We are just returning to the .NET 9 behavior here.

Fixes #117717. May also fix some of the other regressions that have
not yet been analyzed in depth.