Use compiler-lowered DAM when trimming (#117624)
This uses compiler-lowered DAM for trimming compiler-generated code,
and adds test infrastructure and coverage for
DynamicallyAccessedMembers that is lowered by us for the compiler.

We also want to use the compiler-lowered attributes for generated
state machines, when trimming a .NET10+ assembly (leaving the old
heuristics for .NET9 and below, to avoid breaking trim analysis for
existing assemblies). However, the tests uncovered an issue with the
Roslyn implementation:
https://github.com/dotnet/roslyn/issues/79333. Until that is fixed, we
don't rely on the lowered attribute for mapping type parameters by
default (the new behavior is opt-in under a new flag). Once we get a
fix we can remove this.

The old heuristics are still run for all TFMs simply for coverage, but
we don't use the result.

This adds test infrastructure to let us test against a polyfilled DAM
with CompilerLoweringPreserve (since we don't run ILLink/ILC tests
against a live corelib).

The existing ILLink tests (including those shared with ILC trimming
tests) are now built with TargetFrameworkAttribute marking the
assembly as targeting .NET 10. The compiler-generated code tests have
been duplicated for .NET 9 and .NET 10 to test both the old and new
behaviors.

Also fixes a test infrastructure issue where in a "copy" assembly, the
DAM and CompilerLoweringPreserve types that are compiled in are kept,
but the test infrastructure was filtering these out of the
"linkedMembers", leading to validation errors.