Teach regex about more disjoint sets
The regex optimizer tries to make loops atomic whenever it's safe to do so, as doing so removes possible backtracking. When it finds a set loop followed by another set, to be able to make the loop atomic, it needs to prove that the sets are disjoint, that there's nothing that can match both sets. It already does so for a wide variety of cases, but it's missing logic today that helps it to determine that for sets composed only of Unicode categories, e.g. \d, \w, p{C}, etc. This adds that in.