Fix ILLink ref assembly build breaks (#117875)
This is blocking https://github.com/dotnet/dotnet/pull/1546. Looks like this got lost in all the WASM timeout noise in #117624.

We can either suppress or add to ref assembly, but ref assembly only has things ILLink custom steps depend on. So suppression it is.

I just regenerated it with `/p:ApiCompatGenerateSuppressionFile=true` so it shuffled two unrelated lines.