[STJ] Optimize  EnumFieldInfo sorting (#117839)
* [STJ] Only compute PopCount once when topologically sorting Enums

Packs the calculated (once per entry) PopCount into the high 32 bits of the long and the original index into the low 32 bits. Then that can just be sorted using the heavily optimized Array.Sort() method.
After sorting just extract the low 32 bits as the original array index. As before, we negate the actual _PopCount_ to ensure that `Key`s with more on-bits (e.g. more flags represented) will sort **first**.

This trades 2 x O(N log N) [average case] to 2 x O(N^2) [worst case] calls to the `popcount` instruction (or the emulation if NET is not defined) for N **shift-left-32** and **or** + N x **truncate to 32 bits**. It _also_ eliminates the overhead of the `CompareTo` method as it's now a direct `long` low-level compare.

* PR Feedback

Switched to Tuple

* PR Feedback
Switched to native tuple syntax
Co-authored-by: Pranav Senthilnathan <pranav.senthilnathan@live.com>

* Update src/libraries/System.Text.Json/src/System/Text/Json/Serialization/Converters/Value/EnumConverter.cs

---------

Co-authored-by: Eirik Tsarpalis <eirik.tsarpalis@gmail.com>