[EventPipe] Directly store thread_id in seq point map
EventPipeSequencePoints store all alive EventPipeThreads IDs and their
corresponding latest read sequence number in a map which is ultimately
used to write the sequence point block. Previously, the native thread ID
was dynamically retrieved through the EventPipeThreadSessionState,
requiring ref counting the EventPipeThread so that it would not be freed
before the immutable native thread id was written. To simplify, directly
map the native thread ids and remove associated ref counting.