[mono][aot] Prevent localloc in a loop during constrained gsharedvt calls (#117679)
* [mono][aot] Prevent localloc in a loop during constrained gsharedvt calls

We create a var that stores the address of some localloc memory. This var is nulled at method entry. In every place where we need to obtain temporary localloc memory, we check if the cache var was initialized, if not we do a localloc, otherwise we use the cached ptr as the temporary memory buffer. This commit adds 2 such caches because the constrained gsharedvt call can require 2 separate temporary buffers.

* Re-enable test

* milos review